requirements.txt
flask==3.0.0
requests==2.31.0
Werkzeug>=2.0.0,<3.1.0
sentence-transformers==2.2.2
huggingface-hub==0.17.3
torch>=2.0.0
numpy>=1.20.0
python-dotenv==1.0.0
logging==0.5.1
